Move enum to shared types
